another try, GCC is not complaining..

+++ src/lua_api/actor_wrapper.cpp
@@ -8,7 +8,7 @@ namespace amarlon { namespace lua_api {
 
 ActorWrapper::ActorWrapper(ActorPtr actor, const char* name)
 {
-  _actor = ((unsigned)actor.get());
+  _actor = ((unsigned long long)actor.get());
   LuaState& state = Engine::instance().getLuaState();
   state[name].SetObj<lua_api::ActorWrapper>( *this,
                                              "getLevel", &lua_api::ActorWrapper::getLevel);
